# openapi/coinbase-admin-api.yaml
openapi: 3.1.0
info:
  title: Coinbase Admin API
  description: Enterprise API for the Coinbase trading engine. Read-only operator
    routes are active; mutating HTTP routes use shared command services but remain
    live-disabled by the approval gate.
  version: 0.1.0
paths:
  /api/v1/admin/bootstrap:
    get:
      tags:
      - admin
      summary: Read backend association and live-action posture
      operationId: admin_bootstrap_api_v1_admin_bootstrap_get
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBootstrapResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/session:
    get:
      tags:
      - admin
      summary: Read authenticated actor and RBAC evidence
      operationId: admin_session_api_v1_admin_session_get
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSessionResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/capabilities:
    get:
      tags:
      - admin
      summary: Read backend-owned Admin API capability registry
      operationId: admin_capabilities_api_v1_admin_capabilities_get
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCapabilityRegistryResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/release-gate:
    get:
      tags:
      - admin
      summary: Read release-gate posture without running tests from the browser
      operationId: admin_release_gate_api_v1_admin_release_gate_get
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGateReadResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/recovery-gate:
    get:
      tags:
      - admin
      summary: Read recovery readiness posture
      operationId: admin_recovery_gate_api_v1_admin_recovery_gate_get
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGateReadResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/frontend-fixtures:
    get:
      tags:
      - admin
      summary: Read backend-owned fixtures for frontend mock synchronization
      operationId: admin_frontend_fixtures_api_v1_admin_frontend_fixtures_get
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFrontendFixturesResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminApiActionClass:
      type: string
      enum:
      - read_only
      - local_state_mutation
      - live_exchange_place
      - live_exchange_cancel
      - admin_runtime
      - audit
      title: AdminApiActionClass
      description: Enterprise Admin API route risk/action classification.
    AdminApiActor:
      properties:
        actor_id:
          type: string
          minLength: 1
          title: Actor Id
        roles:
          items:
            $ref: '#/components/schemas/AdminApiRole'
          type: array
          title: Roles
      additionalProperties: false
      type: object
      required:
      - actor_id
      title: AdminApiActor
      description: Authenticated actor evidence supplied by future auth middleware.
...
    AdminApiErrorCode:
      type: string
      enum:
      - auth_required
      - permission_denied
      - validation_error
      - idempotency_conflict
      - guard_blocked
      - not_implemented
      - backend_unavailable
      - request_error
      title: AdminApiErrorCode
      description: Structured Admin API error codes exposed to the frontend.
    AdminApiErrorResponse:
      properties:
        code:
          $ref: '#/components/schemas/AdminApiErrorCode'
        message:
          type: string
          title: Message
        severity:
          $ref: '#/components/schemas/AdminApiErrorSeverity'
        guard_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Guard Name
        field_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Path
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        audit_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Audit Id
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - severity
      title: AdminApiErrorResponse
      description: Structured error body shared by Admin API routes.
    AdminApiErrorSeverity:
      type: string
      enum:
      - info
      - warning
      - error
      - critical
      title: AdminApiErrorSeverity
      description: Operator-facing severity for structured Admin API errors.
    AdminApiGateStatus:
      type: string
      enum:
      - passed
      - warning
      - blocked
      - not_applicable
      title: AdminApiGateStatus
      description: Release and recovery gate status exposed by read-only Admin API
        routes.
...
    AdminApiPermission:
      type: string
      enum:
      - analytics:read
      - audit:read
      - order:create
      - order:cancel
      - campaign:read
      - campaign:execute
      - config:update
      - runtime:pause
      - runtime:resume
      - runtime:shutdown
      title: AdminApiPermission
      description: Backend-enforced Admin API permission names.
    AdminApiRole:
      type: string
      enum:
      - viewer
      - operator
      - trader
      - admin
      - auditor
      - emergency
      title: AdminApiRole
      description: Backend-recognized Admin API role names.
    AdminApiRouteAvailability:
      type: string
      enum:
      - available
      - live_disabled
      - contract_pending
      - backend_blocked
      title: AdminApiRouteAvailability
      description: Backend availability posture for frontend-visible routes.
...
    AdminApiSessionStatus:
      type: string
      enum:
      - signed_in
      - signed_out
      - expired
      - forbidden
      title: AdminApiSessionStatus
      description: Authenticated Admin API session status.
    AdminBootstrapResponse:
      properties:
        type:
          type: string
          title: Type
          default: admin_bootstrap
        backend_repository:
          type: string
          title: Backend Repository
        api_version:
          type: string
          title: Api Version
        schema_version:
          type: string
          title: Schema Version
        environment:
          type: string
          title: Environment
        mutating_routes_live_disabled:
          type: boolean
          title: Mutating Routes Live Disabled
        live_execution_enabled:
          type: boolean
          title: Live Execution Enabled
        auth_required:
          type: boolean
          title: Auth Required
        cors_configured:
          type: boolean
          title: Cors Configured
        csrf_required:
          type: boolean
          title: Csrf Required
        capabilities_route:
          type: string
          title: Capabilities Route
        session_route:
          type: string
          title: Session Route
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - backend_repository
      - api_version
      - schema_version
      - environment
      - mutating_routes_live_disabled
      - live_execution_enabled
      - auth_required
      - cors_configured
      - csrf_required
      - capabilities_route
      - session_route
      title: AdminBootstrapResponse
      description: Frontend bootstrap payload for backend association and safety posture.
    AdminCapabilityItem:
      properties:
        route:
          type: string
          title: Route
        method:
          type: string
          title: Method
        action_class:
          $ref: '#/components/schemas/AdminApiActionClass'
        permission:
          anyOf:
          - $ref: '#/components/schemas/AdminApiPermission'
          - type: string
          title: Permission
        availability:
          $ref: '#/components/schemas/AdminApiRouteAvailability'
        live_enabled:
          type: boolean
          title: Live Enabled
        frontend_safe:
          type: boolean
          title: Frontend Safe
        shared_method:
          type: string
          title: Shared Method
        notes:
          type: string
          title: Notes
      additionalProperties: false
      type: object
      required:
      - route
      - method
      - action_class
      - permission
      - availability
      - live_enabled
      - frontend_safe
      - shared_method
      - notes
      title: AdminCapabilityItem
      description: One backend-owned capability advertised to the frontend.
    AdminCapabilityRegistryResponse:
      properties:
        type:
          type: string
          title: Type
          default: admin_capabilities
        capabilities:
          items:
            $ref: '#/components/schemas/AdminCapabilityItem'
          type: array
          title: Capabilities
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      title: AdminCapabilityRegistryResponse
      description: Route/capability registry for frontend navigation and diagnostics.
    AdminFrontendFixturesResponse:
      properties:
        type:
          type: string
          title: Type
          default: admin_frontend_fixtures
        schema_version:
          type: string
          title: Schema Version
        fixtures:
          type: object
          title: Fixtures
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: true
      type: object
      required:
      - schema_version
      title: AdminFrontendFixturesResponse
      description: Backend-owned fixture examples for frontend mock synchronization.
    AdminGateCheck:
      properties:
        name:
          type: string
          title: Name
        status:
          $ref: '#/components/schemas/AdminApiGateStatus'
        detail:
          type: string
          title: Detail
      additionalProperties: false
      type: object
      required:
      - name
      - status
      - detail
      title: AdminGateCheck
      description: One release/recovery check exposed to the frontend.
    AdminGateReadResponse:
      properties:
        type:
          type: string
          title: Type
        status:
          $ref: '#/components/schemas/AdminApiGateStatus'
        checks:
          items:
            $ref: '#/components/schemas/AdminGateCheck'
          type: array
          title: Checks
        read_only:
          type: boolean
          title: Read Only
          default: true
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - type
      - status
      title: AdminGateReadResponse
      description: Read-only gate response for release and recovery views.
...
    AdminSessionResponse:
      properties:
        type:
          type: string
          title: Type
          default: admin_session
        status:
          $ref: '#/components/schemas/AdminApiSessionStatus'
        actor:
          $ref: '#/components/schemas/AdminApiActor'
        permissions:
          items:
            $ref: '#/components/schemas/AdminApiPermission'
          type: array
          title: Permissions
        bearer_token_visible_to_browser:
          type: boolean
          title: Bearer Token Visible To Browser
          default: false
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - status
      - actor
      title: AdminSessionResponse
      description: Authenticated session/RBAC evidence for the frontend.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError